Portal (Wormhole) bridge adapter rework and fixes
The Portal token bridge calls into the Wormhole core contract when tokens are
deposited or transferred to another chain.  The core contract emits a
`LogMessagePublished` event with the Portal token bridge as the sender argument, we can filter on
this to get the list of TXs where tokens were deposited. The previous adapter
wasn’t including deposits made through internal transactions, this is now fixed.
Until the Portal token bridge is upgraded to include the `TransferRedeemed` event,
we rely on the Etherscan API to fetch TXs where tokens were withdrawn.