Portal - EVM wrapped tokens 'to' address change
Wormhole wrapped tokens are burned by the bridge when transferred out from a
chain, thus the `to` argument in the wrapped token contract's `Transfer` event is the
EVM zero-address. However wormhole wrapped tokens are never double-wrapped, so
they should be included in the volume when they're not being sent back to their
origination chain where they'd be included in the withdrawal volume when unlocked. This
change is consistent with DefiLlama's volume calculation methodology of not double
counting token transfers in the volume.